fix: address kt3k's comments
