+++ streams/to_array_buffer_test.ts
@@ -1,9 +1,9 @@
 // Copyright 2018-2023 the Deno authors. All rights reserved. MIT license.
 
 import { assertEquals } from "../assert/assert_equals.ts";
-import { arrayBufferFromReadableStream } from "./array_buffer_from_readable_stream.ts";
+import { toArrayBuffer } from "./to_array_buffer.ts";
 
-Deno.test("[streams] arrayBufferFromReadableStream", async () => {
+Deno.test("[streams] toArrayBuffer", async () => {
   const stream = new ReadableStream<Uint8Array>({
     start(controller) {
       controller.enqueue(Uint8Array.of(1, 2, 3, 4, 5));
@@ -13,6 +13,6 @@ Deno.test("[streams] arrayBufferFromReadableStream", async () => {
     },
   });
 
-  const buf = await arrayBufferFromReadableStream(stream);
+  const buf = await toArrayBuffer(stream);
   assertEquals(buf, Uint8Array.of(1, 2, 3, 4, 5, 6, 7, 8, 9).buffer);
 });
